refactor: move route markup into screens
